refactored methods into the base class
